geaux_tor/onion_scraper/requirements.txt

scrapy>=2.11.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pysocks>=1.7.1